% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_species_in_all_countries.R
\name{fetch_species_for_countries}
\alias{fetch_species_for_countries}
\title{Fetch plant species data for each country in the provided list}
\usage{
fetch_species_for_countries(country_list)
}
\arguments{
\item{country_list}{A character vector of country names.}
}
\value{
A data frame with columns:
\item{country}{The name of the country.}
\item{scrubbed_species_binomial}{The scientific name of each species found.}
}
\description{
This function retrieves all plant species for each country in the specified list
using the BIEN database.
}
\examples{
\dontrun{
countries <- c("Canada", "Brazil")
species_data <- fetch_species_for_countries(countries)
}
}
