% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_circadian_map.R,
%   R/plot_species_count_map.R
\name{plot_species_count_map}
\alias{plot_species_count_map}
\title{Plot Dominant Circadian Pattern by Country on a World Map}
\usage{
plot_species_count_map(country_species_data)

plot_species_count_map(country_species_data)
}
\arguments{
\item{country_species_data}{A data frame with at least two columns:
'country' and 'species'. 'country' should be the name of the
country, and 'species' should be the name of the plant species.}

\item{country_circadian_data}{A data frame with at least two columns:
'country' and 'predominant_clock'. 'country' should be the name of the
country, and 'predominant_clock' should be the circadian pattern.}
}
\value{
A ggplot object displaying a world map with countries colored
by species count.

A ggplot object displaying a world map with countries colored
by species count.
}
\description{
This function takes a data frame with country and circadian information,
and plots it on a world map.
Each country is colored based on nocturnal/diurnal, with blue
indicating nocturnal and yellow indicating diurnal.

This function takes a data frame with country and species information,
aggregates the species count per country, and plots it on a world map.
Each country is colored based on the number of species found, with blue
indicating fewer species and red indicating more species.
}
