% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plant_data.R
\name{read_plant_circadian_data}
\alias{read_plant_circadian_data}
\title{Read plant circadian data from CSV}
\usage{
read_plant_circadian_data(file_path)
}
\arguments{
\item{file_path}{The path to the CSV file with plant circadian data.}
}
\value{
A data frame with columns:
\itemize{
\item \code{scrubbed_species_binomial}: The scientific name of each species.
\item \code{circadian_clock}: The circadian property of each species ("Diurnal" or "Nocturnal").
}.
}
\description{
This function reads a CSV file containing plant species and their circadian clock properties.
}
