% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFloralCircadianR.R
\name{runFloralCircadianR}
\alias{runFloralCircadianR}
\title{Launch Shiny App for FloralCircadianR}
\usage{
runFloralCircadianR()
}
\value{
No return value; launches a Shiny app in the browser.
}
\description{
A function that launches the Shiny app for the FloralCircadianR package.
The app provides a user-friendly interactive interface for alignment and visulization
using functions from the FloralCircadianR package.
}
\examples{
\dontrun{
FloralCircadianR::FloralCircadianR()
}

}
\references{
shiny: Web Application Framework for R. R package version 1.9.1, https://CRAN.R-project.org/package=shiny

Silva, Anjali. TestingPackage. GitHub, https://github.com/anjalisilva/TestingPackage. Accessed 5 Nov. 2024. \href{https://shiny.rstudio.com/}{Link}
}
