% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_interested_country.R
\name{read_interested_countries}
\alias{read_interested_countries}
\title{Read Interested Countries from CSV}
\usage{
read_interested_countries(file_path)
}
\arguments{
\item{file_path}{The path to the CSV file with the list of countries.}
}
\value{
A character vector of country names.
}
\description{
This function reads a CSV file containing a list of countries of interest.
}
\examples{
\dontrun{
country_list <- read_interested_countries("data/test_country_list.csv")
print(country_list)
}
}
